change attributes to nullable

+++ Transbank/Webpay/TransaccionCompleta/Common/DeferredPeriods.cs
@@ -5,12 +5,12 @@ namespace Transbank.Webpay.TransaccionCompleta.Common
     public class DeferredPeriods
     {
         [JsonProperty("amount")]
-        internal int Amount { get; }
+        internal decimal Amount { get; }
         
         [JsonProperty("period")]
         internal int Period { get; }
 
-        public DeferredPeriods(int amount, int period)
+        public DeferredPeriods(decimal amount, int period)
         {
             Amount= amount;
             Period = period;
